# docs/swagger.yaml
basePath: /
definitions:
  lib.DashboardStats:
    properties:
      average_earnings:
        type: integer
      sales_chart:
        items:
          $ref: '#/definitions/lib.SalesStat'
        type: array
      ticket_sales:
        description: By Category/Location
        items:
          $ref: '#/definitions/lib.SalesStat'
        type: array
    type: object
...
  lib.ResetPasswordRequest:
    properties:
      password:
        type: string
      token:
        type: string
    type: object
  lib.Response:
    properties:
      message:
        type: string
      result: {}
      status:
        type: integer
    type: object
  lib.SalesStat:
    properties:
      label:
        type: string
      value:
        type: integer
    type: object
...
  lib.User:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      role_name:
        type: string
    type: object
...
  lib.UserRole:
    properties:
      role_id:
        type: integer
      user:
        $ref: '#/definitions/lib.User'
    type: object
host: localhost:8888
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Movie Ticket Booking System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tikitz API
  version: "1.0"
paths:
  /admin/stats:
    get:
      consumes:
      - application/json
      description: Retrieve sales and ticket statistics for the admin dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.DashboardStats'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Get dashboard statistics
      tags:
      - admin
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send a reset password link to the user's email
      parameters:
      - description: User Email
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/lib.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: User Credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/lib.User'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with token
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User Registration Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/lib.UserRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Register a new user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Verify token and update user password
      parameters:
      - description: Reset Token
        in: query
        name: token
        type: string
      - description: New Password Details
        in: body
        name: reset
        required: true
        schema:
          $ref: '#/definitions/lib.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Reset password
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"